// inventory/src/app/stocks/models/product.model.ts
export interface Product {
  _id: string;
  productId: string;
  category: string;
  group: string;
  productName: string;
  model: string;
  productDescription: string;
  productPrice: string;
  hasAttributes: string;
  hasInstances: string;
  packSize: string;
  averageCost: string;
  singleUnitProductCode: string;
  measurementUnit: string;
  dimensionGroup: string;
  warrentyTerms: string;
  productQuantity: string;
  productSellingPrice: string;
  productCostPrice: string;

}